Move char2HexInt to com.facebook.presto.teradata.functions.TeradataStringFunctions
